Add byte[] bypass methods to JsonGenerator
Allows users with some known combination of byte[] and Charset to avoid unnecessary copies if the output generator supports doing so. Provides an easy API in other cases.

These methods take an explicit java.nio.Charset instead of a JsonEncoding or attempting autodetection to allow users to pass in fragments or encodings that compose/transcode to valid json but are not json on their own.