JsonWriteContext#writeFieldName not updating _currentName correctly
Currently, for a JSON like this -

{
"a": {
},
"b": {
"c": "cval"
}
}

When writeFieldName is called for node 'b', the _currentName is not updated
as the _gotName value is already set to true (because of the earlier processing
done for 'a'). This means that later down the line, when I try to do a
context.getParent().getCurrentName(), I get 'a' back, instead of 'b'.

This PR tentatively fixes that by pulling the setting of the _currentName earlier
before the function exits.