Issue #455 getCurrentLocation() reports location past end of string at end of input
Fixes the getCurrentLocation() issue by moving changes to _currInputProcessed
and _currInputRowStart inside successful branch. Also includes checks for
getTokenLocation() that make sure correct beginning location of tokens is
reported and that getCurrentLocation() correctly reports the end of
stream when reading is finished.

Tests would be much nicer with parameterized tests from Junit5 since it's
hard to tell which test case fails if one does. That's what the commented
out print statements are for. Let me know if you'd like something different,
but I originally had them separated out and forgot to add methods when I
added to the enum.

Also, the idea for the enum and much of the test data is from
https://github.com/leadpony/jsonp-test-suite, which is also Apache 2
licensed. I had to rework the expected results because the JsonP
getLocation() works very differently. Let me know if you need a
CLA from the author of that project and I'll approach him.