Second attempt to move the spacing config to Separators.
This one is much cleaner than the first attempt but we have to decide which PrettyPrinter config to break: Either the default Separators is with spaces around the colon, then MinimalPrettyPrinter has to modify the default. Or the default is without spaces, then DefaultPrettyPrinter has to modify it.

Of course, this affects anyone who tries to build their own PrettyPrinter using the default Separators.